+++ test/expression/node/FunctionAssignmentNode.test.js
@@ -6,6 +6,7 @@ var assert = require('assert'),
     ConstantNode = require('../../../lib/expression/node/ConstantNode'),
     OperatorNode = require('../../../lib/expression/node/OperatorNode'),
     FunctionAssignmentNode = require('../../../lib/expression/node/FunctionAssignmentNode'),
+    AssignmentNode = require('../../../lib/expression/node/AssignmentNode'),
     RangeNode = require('../../../lib/expression/node/RangeNode'),
     SymbolNode = require('../../../lib/expression/node/SymbolNode');
 
@@ -178,6 +179,15 @@ describe('FunctionAssignmentNode', function() {
     assert.equal(n.toString(), 'function f(x) = 2 + x');
   });
 
+  it ('should stringify a FunctionAssignmentNode conataining an AssignmentNode', function () {
+    var a = new ConstantNode(2);
+
+    var n1 = new AssignmentNode('a', a);
+    var n = new FunctionAssignmentNode('f', ['x'], n1);
+
+    assert.equal(n.toString(), 'function f(x) = (a = 2)');
+  });
+
   it ('should LaTeX a FunctionAssignmentNode', function() {
     var a = new ConstantNode(2);
     var x = new SymbolNode('x');
@@ -185,7 +195,15 @@ describe('FunctionAssignmentNode', function() {
     var p = new OperatorNode('^', 'pow', [o, a]);
     var n = new FunctionAssignmentNode('f', ['x'], p);
 
-    assert.equal(n.toTex(), 'f\\left({x}\\right)={\\left({\\frac{x}{2}}\\right)^{2}}');
+    assert.equal(n.toTex(), 'f\\left({x}\\right)={\\left({\\frac{x}{2}}\\right) ^ {2}}');
   });
 
+  it ('should LaTeX a FunctionAssignmentNode containing an AssignmentNode', function () {
+    var a = new ConstantNode(2);
+
+    var n1 = new AssignmentNode('a', a);
+    var n = new FunctionAssignmentNode('f', ['x'], n1);
+
+    assert.equal(n.toTex(), 'f\\left({x}\\right)=\\left({{a}={2}}\\right)');
+  });
 });