+++ test/expression/node/FunctionNode.test.js
@@ -278,7 +278,14 @@ describe('FunctionNode', function() {
 
     var o = new OperatorNode('+', 'add', [c1, c2]);
     var n3 = new FunctionNode('permutations', [o]);
-    assert.equal(n3.toTex(), '{\\left({4}+{5}\\right)!}');
+    assert.equal(n3.toTex(), '{\\left({4} + {5}\\right)!}');
+  });
+
+  it ('should have an identifier', function () {
+    var a = new ConstantNode(2);
+    var n = new FunctionNode('factorial', [a]);
+
+    assert.equal(n.getIdentifier(), 'FunctionNode:factorial');
   });
 
 });